refactor(server): extract CLI entrypoint bundle to cli.py
Task H (final) of docs/current/server-py-module-split.md. Pure
extraction, verbatim move verified byte-identical against git history
(run_server_with_shutdown, create_server, apply_runtime_preset, main --
192 lines). The `if __name__ == "__main__":` guard stays in server.py
per spec, now calling the re-exported `main`.

server.py keeps a bottom-of-file re-export (`from .cli import main`)
so the pip console script (marm_mcp_server.server:main, pyproject.toml
line 56) keeps resolving without a packaging-metadata change, per the
spec's explicit guidance.

Deviation from the spec's stated approach, found and fixed during
validation: the spec proposed `cli.py` importing `app` from `.server`
at module level. That works when `server` happens to import first
(console script, __main__.py, and the test harness's load_isolated_server
all do this), but genuinely breaks -- not just in theory, it broke 4
real tests (test_compaction_worker.py's 3 apply_runtime_preset tests
plus my own manual check) -- when anything imports `marm_mcp_server.cli`
as the first touch of the package in a process, since server.py's
bottom-of-file `from .cli import main` then tries to import a cli
module that's still mid-import waiting on server. Fixed by moving
`from .server import app` from cli.py's module scope into the two
function bodies that actually use it (run_server_with_shutdown,
create_server) -- app is never referenced at cli.py module scope, so
this fully decouples the import-time cycle in both directions without
changing any logic. Diffed the two functions against the original
verbatim-verified slice to confirm only these two added import lines
differ from the extracted body. Verified both import orders
(`import server` first, `import cli` first) now succeed.

Removed 12 now-dead imports from server.py (sys, typing.Optional,
uvicorn, config.settings bare import, SCHEDULER_AVAILABLE,
SEMANTIC_SEARCH_AVAILABLE, SERVER_HOST, SERVER_PORT, core.memory as
memory_module, core.rate_limiter.rate_limiter,
utils.dependency_check.check_dependencies,
utils.security.generate_api_key) -- all exclusively used by the
functions that just moved; ANALYTICS_DB_PATH/DEFAULT_DB_PATH/
SERVER_VERSION correctly remain (still used by lifespan).

Two test files reference the moved functions without a live re-export
(same category as Tasks E/F, but here as a plain AttributeError since
apply_runtime_preset was never module-global-coupled the way the
tracker/scheduler were -- it only mutates the real settings/memory_module/
rate_limiter objects it's given, so calling it via any binding is
correct): test_cli_entrypoint.py's 3 tests calling
server.apply_runtime_preset(...), and test_compaction_worker.py's 3
tests importing `from marm_mcp_server.server import apply_runtime_preset`.
Repointed both to marm_mcp_server.cli; no assertion or test logic
changed. In test_cli_entrypoint.py, also dropped 2 now-genuinely-unused
`server =` bindings (ruff F841) where server was only ever used for the
now-repointed call.

Functional verification: full local test suite run (590 passed, 13
skipped for genuinely unavailable deps, 0 failed) including
test_cli_entrypoint.py's real subprocess tests that invoke
`python -m marm_mcp_server --generate-key`, `--check-deps`, plain
`import marm_mcp_server`, and a full server boot + /health round trip
through the actual console-script/__main__.py/server.py/cli.py chain --
not mocked. create_server() verified to return the real live app
instance. server.py is now 155 lines (spec estimated ~200-250; lower
because several of the extracted units in B-H carried more of their
own dependency wiring out with them than the spec's estimate assumed).
